Splitt hent opphørsperioder ut av VedtaksperiodeService
